fix: 주식 수 정수 -> 유리수 dto 변경

+++ ET-user/src/main/java/com/team1/etuser/user/dto/SettlementDTO.java
@@ -18,7 +18,7 @@ public class SettlementDTO {
     private String stockCode;
     private Position position;
     private BigDecimal orderPrice;
-    private int orderAmount;
+    private BigDecimal orderAmount;
     private Position orderPosition;
 
     @Override
